move unsafe code to submodule
